+++ core/template_bundle/template_codec/compile_options.h
@@ -112,6 +112,7 @@ struct CompileOptions {
   bool enable_parse_int_flex_{false};
   bool enable_flex_basis_zero_percent_{false};
   bool enable_grid_placement_shorthands_{false};
+  bool enable_grid_lanes_{false};
   bool enable_css_rule_{false};
 };