+++ js_libraries/type-config/config-keys.js
@@ -21,6 +21,7 @@ const configKeys = [
   'enableFetchAPIStandardStreaming',
   'enableFixedNew',
   'enableFlexBasisZeroPercent',
+  'enableGridLanes',
   'enableGridPlacementShorthands',
   'enableJsBindingApiThrowException',
   'enableLayoutOnlyEventThrough',